+++ packages/runtime-v2/plugin-runtime/package.json
@@ -200,6 +200,7 @@
     "@scripts/build": "workspace:*",
     "@scripts/jest-config": "workspace:*",
     "@testing-library/react": "^13.4.0",
+    "@types/loadable__server": "5.12.11",
     "@types/invariant": "^2.2.30",
     "@types/jest": "^29",
     "@types/loadable__webpack-plugin": "^5.7.3",
