+++ packages/runtime-v2/plugin-runtime/src/core/server/index.node.ts
@@ -2,4 +2,12 @@ export { renderString } from './string';
 
 export { renderStreaming } from './stream/index.node';
 
+export {
+  createRequestHandler,
+  type HandleRequest,
+  type HandleRequestOptions,
+  type HandleRequestConfig,
+  type CreateRequestHandler,
+} from './requestHandler';
+
 export type { RenderStreaming, RenderString, RenderOptions } from './shared';
