+++ packages/runtime-v2/plugin-runtime/src/core/server/stream/createReadableStream.node.ts
@@ -3,7 +3,7 @@ import { renderToPipeableStream } from 'react-dom/server';
 import { createReadableStreamFromReadable } from '@modern-js/runtime-utils/node';
 import { ServerStyleSheet } from 'styled-components';
 import { RenderLevel } from '../shared';
-import { ESCAPED_SHELL_STREAM_END_MARK } from '../../common';
+import { ESCAPED_SHELL_STREAM_END_MARK } from '../../../common';
 import { getTemplates } from './template';
 import {
   CreateReadableStreamFromElement,
