+++ packages/runtime-v2/plugin-runtime/src/core/server/stream/shared.ts
@@ -1,6 +1,6 @@
 import { time } from '@modern-js/runtime-utils/time';
 import { createElement } from 'react';
-import { RuntimeContext } from '../../core/context';
+import { RuntimeContext } from '../../context';
 import { HandleRequestConfig } from '../requestHandler';
 import { RenderStreaming } from '../shared';
 import {
