+++ packages/runtime-v2/plugin-runtime/src/core/server/stream/template.ts
