+++ packages/runtime-v2/plugin-runtime/src/core/server/string/prefetch.tsx
@@ -2,8 +2,8 @@ import { renderToStaticMarkup } from 'react-dom/server';
 import { run } from '@modern-js/runtime-utils/node';
 import { ChunkExtractor } from '@loadable/server';
 import { time } from '@modern-js/runtime-utils/time';
-import { RuntimeContext } from '../../core/context';
-import { LoaderResult } from '../../core/loader/loaderManager';
+import { RuntimeContext } from '../../context';
+import { LoaderResult } from '../../loader/loaderManager';
 import { HandleRequestConfig } from '../requestHandler';
 import { SSRErrors, SSRTimings, Tracer } from '../tracer';
 // import { SSRPluginConfig } from '';
